rewrite tokenizer in wasm
* tried my best to keep all comments
* tried my best to keep the code structure similiar with the original JS code

There are still some ways to improve the wasm tokenizer.
I'm not sure in the original code `charCodeAt` will or won't exceed the length of the `source`.  If the answer is not, some of the boundary checking code could be removed.

If this WASM impl could be accepted, could you please publish the C code in the npm tarball?  Therefore we could easily intergrate the tokenizer with our other part of C code.

The reason why I rewrite the tokenizer is that we have to transform the style string on runtime in Browser for our cross platform rendering freamwork(https://github.com/lynx-family). Although the current css-tree's tokenizer is really fast, it still becomes one of a bottelneck of the rendering performance.